feat: moves proxy helper to http class
