+++ Core/Core/Api/Helpers.cs
@@ -13,6 +13,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Sentry;
 using Speckle.Core.Credentials;
+using Speckle.Core.Helpers;
 using Speckle.Core.Kits;
 using Speckle.Core.Logging;
 using Speckle.Core.Models;
@@ -176,7 +177,7 @@ namespace Speckle.Core.Api
 
       try
       {
-        HttpClient client = Helpers.GetHttpProxyClient();
+        HttpClient client = Http.GetHttpProxyClient();
         var response = await client.GetStringAsync($"{_feedsEndpoint}/{slug}.json");
         var connector = JsonSerializer.Deserialize<Connector>(response);
 
@@ -288,83 +289,6 @@ namespace Speckle.Core.Api
       Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData, Environment.SpecialFolderOption.Create);
 
 
-    /// <summary>
-    /// Checks if the user has a valid internet connection
-    /// </summary>
-    /// <returns>True if the user is connected to the internet, false otherwise.</returns>
-    public static async Task<bool> UserHasInternet()
-    {
-      //can ping cloudfare, skip further checks
-      //this method should be the fastest
-      if (await Ping("1.1.1.1"))
-        return true;
 
-
-      //lastly, try getting the default Speckle server, in case this is a sandboxed environment
-      return await HttpPing(AccountManager.GetDefaultServerUrl());
-    }
-
-    /// <summary>
-    /// Pings a specific url to verify it's accessible.
-    /// </summary>
-    /// <param name="hostnameOrAddress">The hostname or address to ping.</param>
-    /// <returns>True if the the status code is 200, false otherwise.</returns>
-    public static async Task<bool> Ping(string hostnameOrAddress)
-    {
-      try
-      {
-        Ping myPing = new Ping();
-        var hostname = (Uri.CheckHostName(hostnameOrAddress) != UriHostNameType.Unknown) ? hostnameOrAddress : (new Uri(hostnameOrAddress)).DnsSafeHost;
-        byte[] buffer = new byte[32];
-        int timeout = 1000;
-        PingOptions pingOptions = new PingOptions();
-        PingReply reply = myPing.Send(hostname, timeout, buffer, pingOptions);
-        return (reply.Status == IPStatus.Success);
-      }
-      catch (Exception)
-      {
-        return false;
-      }
-    }
-
-    /// <summary>
-    /// Pings and tries gettign data from a specific address to verify it's online.
-    /// </summary>
-    /// <param name="address">The address to use</param>
-    /// <returns>True if the the status code is successful, false otherwise.</returns>
-    public static async Task<bool> HttpPing(string address)
-    {
-      try
-      {
-        HttpClient _httpClient = GetHttpProxyClient();
-
-        _httpClient.Timeout = TimeSpan.FromSeconds(1);
-        var response = await _httpClient.GetAsync(address);
-        return response.IsSuccessStatusCode;
-
-      }
-      catch (Exception)
-      {
-        return false;
-      }
-    }
-
-    public static HttpClient GetHttpProxyClient(HttpClientHandler handler = null)
-    {
-
-      IWebProxy proxy = WebRequest.GetSystemWebProxy();
-      proxy.Credentials = System.Net.CredentialCache.DefaultCredentials;
-
-      if (handler == null)
-      {
-        handler = new HttpClientHandler();
-      }
-      handler.Proxy = proxy;
-      handler.PreAuthenticate = true;
-
-      var client = new HttpClient(handler);
-
-      return client;
-    }
   }
 }
