+++ Core/Core/Logging/Analytics.cs
@@ -12,6 +12,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Web;
 using Speckle.Core.Credentials;
+using Speckle.Core.Helpers;
 using Speckle.Newtonsoft.Json;
 
 namespace Speckle.Core.Logging
@@ -185,7 +186,7 @@ namespace Speckle.Core.Logging
           });
 
           var query = new StreamContent(new MemoryStream(Encoding.UTF8.GetBytes("data=" + HttpUtility.UrlEncode(json))));
-          HttpClient client = Api.Helpers.GetHttpProxyClient();
+          HttpClient client = Http.GetHttpProxyClient();
           client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("text/plain"));
           query.Headers.ContentType = new MediaTypeHeaderValue("application/json");
           client.PostAsync(MixpanelServer + "/track?ip=1", query);
@@ -224,7 +225,7 @@ namespace Speckle.Core.Logging
 
 
           var query = new StreamContent(new MemoryStream(Encoding.UTF8.GetBytes("data=" + HttpUtility.UrlEncode(json))));
-          HttpClient client = Api.Helpers.GetHttpProxyClient();
+          HttpClient client = Http.GetHttpProxyClient();
           client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("text/plain"));
           query.Headers.ContentType = new MediaTypeHeaderValue("application/json");
           client.PostAsync(MixpanelServer + "/engage#profile-union", query);
