+++ DesktopUI2/DesktopUI2/ViewModels/StreamViewModel.cs
@@ -15,6 +15,7 @@ using Material.Icons;
 using Material.Icons.Avalonia;
 using ReactiveUI;
 using Speckle.Core.Api;
+using Speckle.Core.Helpers;
 using Speckle.Core.Kits;
 using Speckle.Core.Logging;
 using Splat;
@@ -1153,7 +1154,7 @@ namespace DesktopUI2.ViewModels
 
         Reset();
 
-        if (!await Helpers.UserHasInternet())
+        if (!await Http.UserHasInternet())
         {
           Dispatcher.UIThread.Post(() =>
             MainUserControl.NotificationManager.Show(new PopUpNotificationViewModel()
@@ -1259,7 +1260,7 @@ namespace DesktopUI2.ViewModels
 
         Reset();
 
-        if (!await Helpers.UserHasInternet())
+        if (!await Http.UserHasInternet())
         {
           Dispatcher.UIThread.Post(() =>
             MainUserControl.NotificationManager.Show(new PopUpNotificationViewModel()
